Tests/LibWeb/Text/expected/ResizeObserver/active-target-survives-earlier-callback-gc.txt

first observer target: first
second observer entries: 1
second observer target: second
first observer callbacks: 1
second observer callbacks: 1
